requirements.txt
streamlit
pandas
sqlalchemy
mysql-connector-python
mysql
python-dotenv
google-generativeai
requests
